Making entire area of “browse files…” button clickable
Making the entire areal of “browse files…” clickable. Before this change, the button was wider than its parent divs, which caused the overflowing parts of the button to lose clickability.
I'm achieving “entire button clickability” by moving this button to another div which can be bigger than the button.

Actually, I'm moving the entire form that contains this button to the new div.
The new li structure looks like this:
<li>
--<div class="flex-li"/>
--<div class="image-upload-inputs">
-----<input type="file">
-----<input type="submit">
--<div>
</li>

the previous structure:
<li>
--<div class="flex-li">
----<div class="image">
------<form class="image-upload">
---------<img src="">
---------<input type="file">
---------<input type="submit">
------</form>
----</div>
--</div>
</li>