Downloading jest
Trying to make saveList function cabable to be unit tested. I'm letting saveList return a text string value 'Fail' or 'Success'. Unit testing a function that uses fetch API isn't that easy though, so I might as well revert this change if I can't manage to make a test for it.