Listing filenames and e2e testing improvements
Improving code for listing filenames of all files within a folder. Before, this was achieved with filenamesInFolder.php, a weird JavaScript&PHP hybrid script. (It also caused an warning in Firefox.)

Adding new steps to the e2e test. Now it also checks for the links presence. (The home page has links to available lists.)