Adding safety checks in tests
When testing for whether elements exists or not, whereas those elements is being added by JavaScript, it might happen that the test is faster than JS is adding the elements.

In these cases, it wrongly indicates that those elements never gets added to the DOM.

I'm making safety checks so that whenever a test doesn't find the element, it will wait a bit and check again.

This way all the tests can run at sloMo: 0